Improve validation of app management files
Verify the existence of sdd_install/sdd_uninstall functions before
sourcing the file and trying to run the function. Otherwise in case of
e.g. installing a valid and an invalid app, the management file of the
first one is sourced, providing an sdd_install function. This function
would then still be present after sourcing the second file (which did
not contain such a function in the first place), and lead to an
apparently successful installation of the second, invalid app.